+++ app/src/main/AndroidManifest.xml
@@ -65,7 +65,7 @@
             android:label="@string/title_activity_register" >
         </activity>
         <activity
-            android:name=".PeriodItemActivity"
+            android:name=".activity.PeriodItemActivity"
             android:label="@string/title_activity_period_item"
             android:theme="@style/half_transparent">
         </activity>
