+++ app/src/main/java/nju/com/piece/activity/PeriodItemActivity.java
@@ -1,17 +1,13 @@
-package nju.com.piece;
+package nju.com.piece.activity;
 
 import android.app.Activity;
 import android.app.TimePickerDialog;
-import android.content.Intent;
-import android.nfc.Tag;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TimePicker;
 
-import java.util.Date;
-
-import nju.com.piece.activity.TimeLineActivity;
+import nju.com.piece.R;
 import nju.com.piece.adapter.adapterEntity.TimelineItem;
 import nju.com.piece.database.DBFacade;
 import nju.com.piece.database.TagType;
