+++ app/src/main/res/layout/activity_period_item.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="fill_parent"
     android:layout_height="fill_parent"
-    tools:context="nju.com.piece.PeriodItemActivity">
+    tools:context="nju.com.piece.activity.PeriodItemActivity">
 
 
     <LinearLayout
